skhuAdventure 2.2 (일부 버그 및 에러 수정버전)
- 테이블명 오타 수정
- 서버포트번호 변경

+++ hj/skhuAdventure/src/main/java/skhu/skhuAdventure/repository/UserRepository.java
@@ -4,10 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import skhu.skhuAdventure.entity.UserEntity;
 
-
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
-//    UserEntity findById(Integer id);
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
     UserEntity findByEmail(String email);
-
 }
